chore: Remove unused code and update button press tracking

+++ PoliwhiRL/utils/utils.py
@@ -50,7 +50,8 @@ def plot_metrics(rewards, losses, episode_steps, button_presses, n=1, save_loc="
     ax2.set_ylabel("Loss")
 
     # Plot button presses as bar chart
-    button_presses = np.array(button_presses)
+    
+    button_presses = np.array(button_presses, dtype=int)
     button_presses = np.bincount(button_presses)
     ax3.bar(actions, button_presses)
     ax3.set_title("Button Presses")
